+++ Telemetry/Pit/backend_debug.py
@@ -1,5 +1,5 @@
 import sys
-from Telemetry.Pit.serial_recv import *
+from Telemetry.serial_recv import *
 
 if len(sys.argv) < 2 or len(sys.argv) > 3:
     print(f"Usage: {sys.argv[0]} <PORT | FILE> [baudrate]")
